+++ src/layman/publication_relation/util.py
@@ -4,6 +4,6 @@ def update_related_publications_after_change(workspace, publication_type, public
     from layman.util import patch_after_feature_change, get_publication_info
 
     if publication_type == LAYER_TYPE:
-        maps = get_publication_info(workspace, publication_type, publication, context={'keys': ['layer_maps']})['_layer_maps']
+        maps = get_publication_info(workspace, publication_type, publication, context={'keys': ['used_in_maps']})['used_in_maps']
         for map_obj in maps:
             patch_after_feature_change(map_obj['workspace'], MAP_TYPE, map_obj['name'])